requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Facebook
facebook-sdk==3.1.0
facebook-scraper==0.2.59

# Gemini
google-generativeai==0.3.2

# Knowledge Base
chromadb==0.4.24
# sentence-transformers==2.2.2

# Rate Limiting
aiolimiter==1.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
tenacity==8.2.3
loguru==0.7.2
pandas==2.1.3
pytz==2023.3.post1

# Google Sheets
gspread==6.0.0
oauth2client==4.1.3
